ReCyclists/pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
discord = "^1.7.3"
Flask = "^2.0.3"
youtube_dl = "^2021.12.17"
bs4 = "^0.0.1"
requests = "^2.27.1"
replit = "^3.2.4"
firebase-admin = "^6.0.0"
emoji = "^2.1.0"
aiohttp = "^3.8.3"
async-timeout = "^4.0.2"
ffmpeg = "^1.4"
pytube = "^12.1.0"
psutil = "^5.9.3"
Pillow = "^9.3.0"
openai = "^0.15.0"
keras = "^2.11.0"
tensorflow = "^2.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"